# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for time tracking.",
        "title": "Time Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "rovsh.dev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "description": "User search with filter and pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "user search with filter and pagination",
                "parameters": [
                    {
                        "description": "search parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.UsersGetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.UsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "create a new user",
                "parameters": [
                    {
                        "description": "series and number of the user's passport",
                        "name": "passport",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.User"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "description": "Get user by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.Response-time_tracker_internal_models_User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "delete a user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.User"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a user data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "update a user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/time_tracker_internal_models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "time_tracker_internal_models.Page": {
            "type": "object",
            "required": [
                "number",
                "size"
            ],
            "properties": {
                "number": {
                    "type": "integer",
                    "minimum": 1
                },
                "size": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "time_tracker_internal_models.Response-time_tracker_internal_models_User": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/time_tracker_internal_models.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "time_tracker_internal_models.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "time_tracker_internal_models.Task": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "time_tracker_internal_models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "task": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/time_tracker_internal_models.Task"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "time_tracker_internal_models.UserCreateRequest": {
            "type": "object",
            "properties": {
                "passport_number": {
                    "type": "string"
                }
            }
        },
        "time_tracker_internal_models.UserUpdateRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "time_tracker_internal_models.UsersGetFilter": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "passport_number": {
                    "type": "string",
                    "maxLength": 11
                },
                "patronymic": {
                    "type": "string",
                    "maxLength": 255
                },
                "surname": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "time_tracker_internal_models.UsersGetRequest": {
            "type": "object",
            "properties": {
                "filter": {
                    "$ref": "#/definitions/time_tracker_internal_models.UsersGetFilter"
                },
                "page": {
                    "$ref": "#/definitions/time_tracker_internal_models.Page"
                }
            }
        },
        "time_tracker_internal_models.UsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/time_tracker_internal_models.User"
                    }
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/time_tracker_internal_models.Page"
                }
            }
        }
    }
}